Authorization

+++ my-app/backend/routes/post.js
@@ -35,7 +35,8 @@ multer({storage:storage}).single("image"),
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   console.log(post);
   post.save().then(result =>{
@@ -94,21 +95,27 @@ multer({storage:storage}).single("image"),
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
   console.log(post);
   console.log(req.params.id)
-  Post.updateOne({_id: req.params.id}, post).then(result=>{
-    res.status(200).json({ message:'updated successful' })
+  Post.updateOne({_id: req.params.id, creator:req.userData.userId}, post).then(result=>{
+    if (result.nModified > 0){
+      res.status(200).json({ message:'updated successful' });
+    }
+      res.status(401).json({message:"Not Authorized!"});
   });
 });
 
 router.delete("/:id",
 checkAuth,
 (req, res, next) =>{
-  Post.deleteOne({_id:req.params.id}).then(result=>{
-    console.log(result);
-    res.status(200).json({ message:"Post deleted! "});
+  Post.deleteOne({_id:req.params.id, creator:req.userData.userId}).then(result=>{
+    if (result.n>0){
+      res.status(200).json({ message:"Post deleted! "});
+    }
+    res.status(401).json({message:"Not Authorized!"});
   });
 });
